refactor: large refactor, see description
- drop prettier
- make eslint & typescript stricter
- simplify Payload & Connection generation
- return error payloads where missing
- prepare support for modrinth
- remove schema.gql from git

+++ api/src/modpack/modpack.resolver.ts
@@ -1,6 +1,5 @@
 import {
   Mutation,
-  ObjectType,
   Resolver,
   Query,
   Field,
@@ -11,14 +10,15 @@ import {
   ResolveField, Parent,
 } from '@nestjs/graphql';
 import { MinLength, MaxLength, IsEnum, IsIn } from 'class-validator';
-import { ModpackService } from './modpack.service';
-import { Modpack } from './modpack.entity';
-import { ModLoader } from '../../../common/modloaders';
 import * as minecraftVersions from '../../../common/minecraft-versions.json';
-import { ModVersion } from '../mod/mod-version.entity';
-import { ModJar, ReleaseType } from '../mod/mod-jar.entity';
+import { ModLoader } from '../../../common/modloaders';
+import type { ModJar } from '../mod/mod-jar.entity';
+import { ReleaseType } from '../mod/mod-jar.entity';
 import { ModService } from '../mod/mod.service';
+import { Payload } from '../utils/graphql-payload';
 import ModpackMod from './modpack-mod.entity';
+import { Modpack } from './modpack.entity';
+import { ModpackService } from './modpack.service';
 
 registerEnumType(ModLoader, {
   name: 'ModLoader',
@@ -28,17 +28,31 @@ registerEnumType(ReleaseType, {
   name: 'ReleaseType',
 });
 
-@ObjectType()
-class CreateModpackPayload {
-  @Field()
-  node: Modpack;
-
-  withNode(node: Modpack): this {
-    this.node = node;
-
-    return this;
-  }
+enum CreateModpackErrorCodes {
 }
+
+const CreateModpackPayload = Payload('CreateModpack', Modpack, CreateModpackErrorCodes);
+
+enum AddModpackModErrorCodes {
+  MODPACK_NOT_FOUND = 'MODPACK_NOT_FOUND',
+}
+
+const AddModpackModPayload = Payload('AddModpackMod', Modpack, AddModpackModErrorCodes);
+
+enum RemoveJarFromModpackErrorCodes {
+  MODPACK_NOT_FOUND = 'MODPACK_NOT_FOUND',
+  JAR_NOT_FOUND = 'JAR_NOT_FOUND',
+}
+
+const RemoveJarFromModpackPayload = Payload('RemoveJarFromModpack', Modpack, RemoveJarFromModpackErrorCodes);
+
+enum ReplaceModpackJarErrorCodes {
+  MODPACK_NOT_FOUND = 'MODPACK_NOT_FOUND',
+  JAR_NOT_FOUND = 'JAR_NOT_FOUND',
+  NEW_JAR_NOT_FOUND = 'JAR_NOT_FOUND',
+}
+
+const ReplaceModpackJarPayload = Payload('ReplaceModpackJar', Modpack, ReplaceModpackJarErrorCodes);
 
 @InputType()
 class CreateModpackInput {
@@ -90,15 +104,15 @@ class RemoveJarFromModpackInput {
 export class ModpackResolver {
 
   constructor(
-    private modpackService: ModpackService,
-    private modService: ModService,
+    private readonly modpackService: ModpackService,
+    private readonly modService: ModService,
   ) {}
 
   @Query(() => Modpack, {
     nullable: true,
     name: 'modpack',
   })
-  getModpack(@Args('id', { type: () => ID }) externalId: string) {
+  async getModpack(@Args('id', { type: () => ID }) externalId: string) {
     return this.modpackService.getModpackByEid(externalId);
   }
 
@@ -106,7 +120,7 @@ export class ModpackResolver {
     nullable: true,
     name: 'modpacks',
   })
-  getModpacks() {
+  async getModpacks() {
     return this.modpackService.getModpacks();
   }
 
@@ -117,45 +131,63 @@ export class ModpackResolver {
     return new CreateModpackPayload().withNode(modpack);
   }
 
-  @Mutation(() => CreateModpackPayload)
-  async addModToModpack(@Args('input') input: AddModpackModInput): Promise<CreateModpackPayload> {
-    const modpack = await this.modpackService.getModpackByEid(input.modpackId);
+  @Mutation(() => AddModpackModPayload)
+  async addModToModpack(@Args('input') input: AddModpackModInput): Promise<typeof AddModpackModPayload.T> {
+    const modpack: Modpack | null = await this.modpackService.getModpackByEid(input.modpackId);
 
-    // TODO: error if modpack = null
+    if (modpack == null) {
+      return new AddModpackModPayload()
+        .withError(AddModpackModErrorCodes.MODPACK_NOT_FOUND, `modpack ${input.modpackId} not found`);
+    }
 
     await this.modpackService.addModUrlsToModpack(modpack, input.byUrl);
 
-    return new CreateModpackPayload().withNode(modpack);
+    return new AddModpackModPayload().withNode(modpack);
   }
 
-  @Mutation(() => CreateModpackPayload)
-  async removeJarFromModpack(@Args('input') input: RemoveJarFromModpackInput): Promise<CreateModpackPayload> {
+  @Mutation(() => RemoveJarFromModpackPayload)
+  async removeJarFromModpack(@Args('input') input: RemoveJarFromModpackInput): Promise<typeof RemoveJarFromModpackPayload.T> {
     const [modpack, jar] = await Promise.all([
       this.modpackService.getModpackByEid(input.modpackId),
       this.modService.getJar(input.jarId),
     ]);
 
-    // TODO: error if modpack = null
-    // TODO: error if jar = null
+    if (modpack == null) {
+      return new RemoveJarFromModpackPayload().withError(RemoveJarFromModpackErrorCodes.MODPACK_NOT_FOUND, `modpack ${input.modpackId} not found`);
+    }
+
+    if (jar == null) {
+      return new RemoveJarFromModpackPayload().withError(RemoveJarFromModpackErrorCodes.JAR_NOT_FOUND, `jar ${input.jarId} not found`);
+    }
 
     await this.modpackService.removeJarFromModpack(modpack, jar);
 
-    return new CreateModpackPayload().withNode(modpack);
+    return new RemoveJarFromModpackPayload().withNode(modpack);
   }
 
-  @Mutation(() => CreateModpackPayload)
-  async replaceModpackJar(@Args('input') input: ReplaceModpackJarInput): Promise<CreateModpackPayload> {
+  @Mutation(() => ReplaceModpackJarPayload)
+  async replaceModpackJar(@Args('input') input: ReplaceModpackJarInput): Promise<typeof ReplaceModpackJarPayload.T> {
     const [modpack, oldJar, newJar] = await Promise.all([
       this.modpackService.getModpackByEid(input.modpackId),
       this.modService.getJar(input.oldJarId),
       this.modService.getJar(input.newJarId),
     ]);
 
-    // TODO: entity missing errors
+    if (modpack == null) {
+      return new ReplaceModpackJarPayload().withError(ReplaceModpackJarErrorCodes.MODPACK_NOT_FOUND, `modpack ${input.modpackId} not found`);
+    }
+
+    if (oldJar == null) {
+      return new ReplaceModpackJarPayload().withError(ReplaceModpackJarErrorCodes.JAR_NOT_FOUND, `jar ${input.oldJarId} not found`);
+    }
+
+    if (newJar == null) {
+      return new ReplaceModpackJarPayload().withError(ReplaceModpackJarErrorCodes.NEW_JAR_NOT_FOUND, `jar ${input.newJarId} not found`);
+    }
 
     await this.modpackService.replaceModpackJar(modpack, oldJar, newJar);
 
-    return new CreateModpackPayload().withNode(modpack);
+    return new ReplaceModpackJarPayload().withNode(modpack);
   }
 
   // TODO: Pagination
